Test works

+++ src/main/java/com/zooplus/fumbliebackend/model/Product.java
@@ -5,7 +5,6 @@ import lombok.Data;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.Id;
-import java.math.BigInteger;
 
 @Entity
 @Data
@@ -17,5 +16,5 @@ public class Product {
     private String title;
     private String description;
     private String image;
-    private BigInteger price;
+    private Double price;
 }
